UsedProvRelation   - behaves like a ProvRelation   - requires relationship_type to be 'used'   - requires relatable_from to be an Activity   - requires relatable_to to be a FileVersion
ModelSharedExamples adds a 'a ProvRelation' shared example for models which inherit from ProvRelation
  expects subject to:
  - validate presence of relationship_type, relable_to, and relatable_from
  - be a kind
  - be a graphed relationship with auto create true
